pkg/highlight: support newlines _inside_ the syntect HTML structure spans, rather than only outside
The new version of syntect_server, which fixes a separate important bug (see
https://github.com/sourcegraph/syntect_server/commit/1ff36cf3a77df80a7559e139b386c043438190ca),
requires that we accept newlines only inside the spans, whereas they used to be
outside of the spans. In the future, we can remove the support for newlines
outside the spans. For now, both are supported and tests are added for both.