frontend: always redirect to appURL first (experimentalFeatures.canonicalURLRedirect) (#11714)
* middleware: make HTTPSRedirect auto-reload config, add tests

This change prepares HTTPSRedirect to be expanded (in a future commit) to support general redirects to the canonical appURL.

There is no change in behavior (except for the case when the httpToHttpsRedirect site config value is invalid).

* frontend: always redirect to appURL first (experimentalFeatures.canonicalURLRedirect)

The new `experimentalFeatures.canonicalURLRedirect` site config property causes the server to force-redirect to the appURL scheme and host if the HTTP request is for a different scheme and host. This is like httpToHttpsRedirect, except that it also forces the correct host.

It removes the JavaScript redirect (`window.location.protocol = ...`) because that is only advantageous when the HTTP response `Location` header is non-absolute (ie contains no host). Now that our `Location` response headers are absolute, reverse proxies will not munge them, so we don't need the JS workaround.

fix #9468