allow node to use over 1.7GB of memory (#11619)
For historical reasons, Node defaults to a strict memory cap of
around 1.7GB, which means that large things like webpack operations
can run out of memory and segfault. This started happening to me
sporadically when running dev/start.sh. This is probably not really
an ideal fix, and it only works for development environments, but it
works for the only case where I've seen the thing.

Note that the failure itself seems to be sporadic, so it's very
hard to prove or disprove that this works.