all: allow usernames and org names of up to 255 chars
GitHub limits usernames to 38 chars, but that's likely for public display purposes (to prevent users from creating annoyingly long usernames). In the on-premises scenario, that is not as important to protect against. GitLab supports names of up to 255 chars, for example.

We may need to reduce this later, esp. for Sourcegraph.com.

fix https://github.com/sourcegraph/issues-twosigma/issues/13