symbols: close ReadCloser for git archive (#11810)
r is a ReadCloser that is never closed. This may cause a memory leak and/or hoarding of HTTP connections. In practice, it is unlikely to, since the reader will be fully read (and the implementation in that case does release the connection and associated resources). If, however, there is an error reading the tar file or the parent context is canceled, the reader might not be fully read, and the resources would be orphaned (and never freed).