xlang/gobuildserver: always log the detected root import path
Prior to this change we did not log the detected root import path. This was
really silly not to do, because root import path detection is often entirely
a heuristic approach that can fail. If we had logged this before, it is likely
that we would have understood the issue of #10395 much sooner (or perhaps even
that the customer would've noticed/fixed it on their own).