xlang/gobuildserver: allow specifying RootImportPath via .sourcegraph/config.json
When I first set out to improve the fact that the "canonical import path detector"
was not good enough (as was the case for a customer in, e.g., #10395) I took
the approach of improving the heuristic first. However, the truth is that while
the heuristic will work in some cases, it is just not possible to make it work
efficiently for all of them. It would drastically increase the initialization
time of large monorepos if we tried to handle all cases (requires reading every
Go file in the repo, period).

Instead, I settled on this approach. When the heuristic is not good enough to
detect your root import path by default, you will see a log line stating it:

```
Detected root import path "code.wrong.com/the/repo" for "git://codestash.slimsag.com/the/repo?ff8e44d960b94bcda49281d0510910499a4c6922"
```

If that root import path is not correct (users would notice this via code intel.
not working on their dependencies, or by spotting the above log line) and will
need to instead inform the Go language server via a new entry to `.sourcegraph/config.json`
in the root of their repo:

```
{
	"go": {
		"RootImportPath": "code.good.com/the/repo"
	}
}
```

The Go LS will detect this and use that repo import path instead of the one
detected (or not detected) by the heuristics we have in place.