backend: use cached repo in GetCommit if repoupdater returns error (#11965)
In cases where the repoupdater reports repo-not-found or repo-unauthorized, this makes GetCommit fall back to using the cached information about the repo. This is how ResolveRev works too (and it was a mistake that GetCommit didn't work this way too) and adds some dumb robustness to code host API errors.

A user who visits a repo that doesn't exist will still see the repo be reported as not found because the `graphql?Repository` call does not ignore the repoupdater error.

Note that this behavior IS technically incorrect. If you have a repo in the DB `foo` and then delete it from the code host, it will continue to be reported as existing on Sourcegraph until an admin manually deletes it. This PR does not affect that behavior.

fix #11957

> This PR updates the CHANGELOG.md file to describe any user-facing changes.