New repo-updater scheduler
This actually implements and supports the feature flag for the new
scheduler. The new and old schedulers can exist simultaneously,
for now. If the feature flag is disabled, the new scheduler should
never get invoked, and should clean itself up nicely. If the feature
flag is enabled, it should run, and everything should get transferred
over to it.