pkg/langservers: cache Info calls (#11995)
* pkg/langservers: cache Info calls

Prior to this change we would often make tons of `langserver.Info` calls. When
visiting the site admin code intelligence page one would be made for each
language every 10s, in addition to one for each language when visiting a repo
of any language due to the code intelligence indicator.

After this change, a background worker polls docker langserver information
every 1s and callers get the latest copy of the information. This makes it much
harder to bog down the system overall, and makes the code intelligence page
load near instantly.

Helps #11989

* CHANGELOG: mention code intelligence page perf

* return neatly