handle-change/go-install: restart things only if they rebuilt cleanly (#11780)
* handle-change/go-install: restart things only if they rebuilt cleanly

When changes are detected, we automatically rebuild and restart affected
services. Make go-install.sh slightly fancier, and make it emit on stdout
only a list of rebuilt things (and only if invoked with -v), with all
diagnostics going to stderr. Then make handle-change.sh invoke go-install
with a list of things to specifically try to rebuild, and restart only
services which were reported as being successfully rebuilt.

The significant change in usage here is that if you make a change to
something in a command, and the change won't compile, you don't see
a successful restart of the command, you see a warning that it failed
to rebuild (and also the compiler errors).

No changelog because customers don't use dev/start.sh, I don't think.

* Clean up raced/notraced logic

The support for -race=true/-race=false makes it possible to slightly
reduce code duplication, I think this is clearer.