suppress pointless ANSI codes from keycloak
keycloak sends each line as "^[[0mTEXT\n^[[0m", resetting ANSI colors
both before and after the line. goreman sees this as two lines, one
of which has no newline, and helpfully displays the second line with a
newline. But the second line has nothing human-readable.

The fancy subshell thing is needed because otherwise, the SIGINT sent
when you ^C goreman kills sed instantly, and keycloak dies because
of writing to a broken pipe. This way, we just get the desired behavior.
Yay.