Watch more narrowly to reduce resource usage
chokidar and gulp both consume a ton of a limited resource,
inotify watch requests. We can't really fix that sanely, but we
can reduce it a bit by explicitly restricting chokidar to the
directories we care about; the difference is about 400 files
(4,000 vs. 4,400 or so) on my system, which represents "directories
somewhere in the tree that it's watching in case someone creates a
.go file in them."

Also, provide an experimental replacement of chokidar for
the handle-change.sh case. This is slightly customized; existing
utilities weren't great for this. We should probably investigate
whether Watchman can do a better job.

chokidar ends up using one inotify-watch slot per *file* in
the directories it watches, and for vendor/*, that's over 2k
files, meaning chokidar is using up >4,000 of the ~8k slots
Linux seems to be configured with by default.

Using inotifywait to do the same thing gets you down to about
900, but there's no debouncing/cleanup of things, and it's a bit
spammy. So what if we used a wrapper?

In process, change handle_change.sh to handle multiple filenames
per invocation, allowing it to coalesce things. Without this
change, modifications to two .go files in the same command
directory could cause the command to be rebuilt twice. Not that
this is a problem with chokidar; it reports at most one file
change per 400ms no matter what, and will simply miss other
changes.