graphqlbackend: fix security note comment
For user settings, there is no separate definition of "can view" and "can update", so this is technically correct now. However, it is safer to document that this is the less strict of the 2, so that if/when we have that distinction in the future, nobody is confused into thinking this performs a stricter access check than it actually does.