graphqlbackend: expose config cascade for all subjects, improve docs
The configuration cascade consists of the final (merged) config for a given configuration subject (i.e., site/org/user) *and* all of its ancestor configs with lower precedence that are merged to produce its final (merged) config.

Also renames the Query.configuration field to Query.viewerConfiguration for consistency (with other fields that implicitly take the viewer as input).