Improve/prettify log15 output
First, move all the log15 initialization into tracer.Init, since
everything that was doing log15 initialization was also calling
tracer.Init(). Also, change Init's calling convention to omit
the name (since it's predictable) and allow specifying a log
filter.

Also improve the logic for creating and installing the log handler;
in particular, setting the handler was contingent on successfully
finding a log level; now, if we don't find the log level,
the log level filter isn't applied but the rest of the code
still works.

Also, add a new environment variable, SRC_LOG_FORMAT, which
allows selecting a log format, one of "logfmt" (the default)
or "condensed" (the new pretty one). The intent is to set this in
dev/start.sh, so we get easier-to-read messages when running on console,
but production environments retain the logfmt display.

(We might want to use a more-logfmt display format, but log15 doesn't
expose its fancy logfmt formatting logic.)