searcher: Do not block store goroutine if search canceled (#12028)
In 820faa99376e89c15a29f4c33732f6ce04687ed2 we changed the way open files from
the store. Previously when the goroutine was cancelled, we would spin up
another goroutine to read from the `resC` chan to ensure we closed the
file. Now that we don't open the file we don't need to ensure we close the
file, but we still need to ensure we read from `resC` to prevent blocking the
store goroutine. Instead of doing that, this fix instead makes the channel
have a buffer of 1, which would prevent the blocked goroutine.