web: factor out the hardcoded filepath-to-mode mapping, allow mode specification by callers
UI components should now call the `getXyz` functions exported in `backend/features`. These add the `mode` (based on file path), and more importantly, by being separate from the LSP request functions, they make it easier in the future to support combining results from multiple LSP servers.

This lets the callers of the lsp module exported functions provide a `mode` property in the args. If present, it is used as the LSP mode. If not present, the old behavior is preserved (the mode is looked up using the file path).

There is no intended behavior change from this commit. It makes it possible (in a future commit) for callers to specify modes that are not directly mapped from the file path (which is important for extensions).