xlang: use jsonrpc2.ObjectStream as interface for conn because it is more specific
This lets us avoid repetitive calls to jsonrpc2.NewBufferedStream when providing the connection to jsonrpc2.NewConn. Every connection used by xlang needs to be convertible to a jsonrpc2.ObjectStream anyway, so there is no loss of expressiveness.

This also makes it easier to (in the future) add support for WebSocket connections.