web: pass through (always-empty) Extensions to all UI components that will need it
This makes it so the `extensions` list is propagated everywhere it will be needed. This list is always empty, as of this commit. The purpose of this commit is to split up the big diff that adds full extension support to get it merged sooner without conflicting.

This commit does not affect the behavior of any component.