make script work if /bin/sh isn't bash (#12071)
Time is a magic builtin in bash, sh doesn't have it, and thus
	time A=b cmd
doesn't work in sh. Switch to "A=b time cmd", which uses the
/usr/bin/time command, which has less-pretty output in this situation,
but this way it works.

Doing this instead of specifying /bin/bash, because it turns out
some systems have no /bin/bash or a very old bash in /bin, and we
don't actually CARE about the time format.