Popover: use full width (fix)
This was intended to make it so alerts in popovers did not cause the popover to be its min-width. That fix unintentionally caused other popovers to not take their full width. There might be a way to solve both, but I couldn't figure it out in a few min. It's acceptable to have popovers with alerts take only their min-width, because the min-width is set to a still-acceptable width.