graphql: add renderMarkdown
This change adds a renderMarkdown method with associated (currently empty) options for rendering the markdown.

I am using this from the frontend for rendering markdown in discussions. I anticipate this will be the only user for a while.

Rendering markdown on the backend in this manner has a few primary benefits:

1. Consistency with how we currently render other Markdown files in the web UI (i.e. same markdown renderer and security guarantees).
2. (not yet implemented) Ability to highlight Markdown code snippets using Syntect and get the same language support & identical theming etc.

I do not anticipate that this would be used to e.g. render hover tooltips on the frontend, as it would introduce a needless round trip. Other use cases can be discussed outside of this change.