Move hover code out of Blob and use on diff pages (#12091)
Moves all the hover code out of `Blob2`.
Every page that shows code can create a "hoverifier" through `createHoverifier()`. It is passed all events it needs to listen to as Observables and a few callbacks, and returns a "hoverifier". The hoverifier can be called to hoverify individual code views, like the blob, or the individual file diffs that are part of a diff. The Hoverifier owns the state so that only one HoverOverlay can be shown on the page at the same time.
The context (repo, rev, file) of a hovered symbol is resolved through a callback so it can differ between blob and diff pages (diffs have multiple files and revs).
The output of the Hoverifier is an Observable that outputs the state needed to render the `HoverOverlay` component.

There is a lot more potential for cleanup here, but this is the first fully functional savepoint! 🎉😍
The `hoverifier` folder would eventually be put into its own npm package to be shared with the browser extension.