keycloak.sh: don't try to use sed unbuffered
Because MacOS utils are apparently much older than the sed
found on FreeBSD, and don't have the -l/-u flags.

This alternative is not quite as robust, but in practice for
the actual data sets we're concerned with, it appears to work.

You might wonder why there's no stripping of the spurious
escape sequences. Well, that's because it doesn't actually
matter; what *matters* is that reading a line of input means
that the output of keycloak.sh gets tossed into a buffer
and waits until a newline, so each trailing escape sequence
with no newline waits in the buffer until the next line happens.
Which is fine. Each line from keycloak's shutdown comes with
one or two useless ANSI reset sequences, but whatever.

(This will lose the last line of text if it's a partial
line; "read i" will, in that case, set the variable but return
a non-successful exit status.)