graphqlbackend: allow all users to list users
This does not reduce security, in fact, because anyone could already request a specific user by ID. By incrementing the ID, anyone could already have requested a list of all users.

This fixes an issue where listing the users of an extension required the user to be a site admin.