redispool: A package for consistent access to redis endpoints (#12203)
* redispool: A package for consistent access to redis endpoints

We have some hacky code in other parts of our codebase which try to re-use
redis store without importing it. Additionally, we want to clean up how we
access redis and do sysreqs. It makes sense to add an easy to use and simple
package.

* redispool: Check if Store is available

Our app can function is Cache is not available, but not if Store is not. So
instead move the sysreq to store.

* redispool: Use consistent envvars for specifying network addresses

This also gives us the ability to just specify REDIS_ENDPOINT if the redis
endpoint is shared (such as in server).

* useractivity: Use redispool.Store

* rcache: Remove unused Keys function