gitserver: Migrate old clones to new style (.git) (#12222)
We have to support clones being in two potential locations `$name` or
`$name/.git`. This migrates us to the `.git` location only. We introduced the
`.git` location to support names which are children of others (some customers
do this).

We introduce a manifest file to track if the migration has run. Going forward
we could track other sorts of rarely changing metadata.