gitserver: Be robust against bad repo dirs in migration (#12236)
* gitserver: Do not fail migration if unexpected error occurs

This shouldn't happen in datacenter, but for users they can do strange things
to the repos directory. Also docker not on linux can do strange things with
permissions. So ensure we are robust against any failures. If we fail to
migrate a directory, it will just cease to exist once we remove the backwards
compatability. Which is fine, since they are likely broken anyways.

* gitserver: Include chmod in migration

We include a best-effort chmod to ensure the repository directories are
RWX. We require them to be, and they are managed by us so this should be the
case already. But we have examples in the wild where they are not.