LSP HTTP gateway: allow 3 (or 4) LSP requests, not just 4, to make server capabilities easier (#12237)
Fix issue reported at https://github.com/sourcegraph/sourcegraph/pull/4045#discussion_r201240216.

Previously we required (in the LSP HTTP gateway) LSP payloads to always have 4 LSP requests:

- initialize
- whatever the actual request is
- shutdown
- exit

This was to be strict about the input to avoid the user violating our stateless abstraction over LSP (eg if they could send 2 actual requests, then they might assume the 2nd actual request would reflect state modifications from the 1st actual request, which we would not actually support).

This usually is OK, but when we just want to fetch server capabilities, we don't need the 2nd request. To work around this previously, we sent a textDocument/didOpen, and to avoid that actually causing changes on the lang server (because some lang servers would then treat the didOpen'd file as being empty), we suppressed those requests in lsp-proxy. To avoid needing this workaround (and simplify the code), this commit allows a 3-request payload consisting of:

- initialize
- shutdown
- exit

> This PR does not need to update the CHANGELOG because it is not user facing