settings schema: don't allow unrecognized props
This fixes an issue where invalid properties could be added to settings (global, org, or user) and no red squiggly would appear to tell the user that the value failed JSON Schema validation.

This prevents mistakes such as one that I just saw, where someone added a Phabricator site config property to global settings on dogfood. It should have a red squiggly because it's ignored (that setting is only valid in site config).