Prevent deadlock from ping operations (#12294)
* Prevent deadlock from ping operations

There exists a potential for deadlock because ping operations
are happening with the mutex, and are channel sends on a channel
the other end of which can be blocked on the mutex, depending on
where in its process updateLoop() is. The easy failure case is
for two pings to be sent in a row; the first one wakes up
updateLoop(), which consumes that first message, then comes
around to the top of its loop.

If, before it gets to its lock, another thing attempts to send
a ping, the thing sending the ping blocks on being able to
send that message, and updateLoop blocks on waiting for its
mutex.

This might seem a hard race to trigger, but it can also happen
very quickly if more than one thing that would cause a ping
happens during the (possibly slow) part of the loop where
updateLoop() is doing the actual processing work. Then it
releases the lock, one thing sends a ping immediately, and another
thing grabs the mutex and starts trying to send a ping.
updateLoop() takes the first ping, and... we're deadlocked.

The underlying fix is to make pings asynchronous, and have them
grab the lock directly (they need it to avoid races on r.ready).
They could also just not check r.ready, but this would increase
the number of pings being processed significantly, when in practice
we never care about more than one ping at a time. I also now
check r.ready to avoid sending redundant pings, which shouldn't
matter much but makes me happier.

Also dropped some extra () from doc comments.