xlang/proxy: allow clients to create isolated LSP session without needing to generate a random session
If a client sends `initializationOptions.session == "cxp"` in the LSP initialize message, the server will generate a random session value for them to ensure that their session will not be "joined" by any other LSP clients (who might otherwise be accessing the same repo and commit ID).

This makes client impls easier. It also gets us part of the way toward having true live sessions in lsp-proxy, where there is a 1-to-1 mapping clientProxyConn <-> serverProxyConn. In that world, lsp-proxy would only be used for implementing xfiles extensions and similar, not for backend connection sharing. That might make sense for code extensions (can discuss further, this commit's value doesn't hinge on that discussion).