xlang: support window/showMessage{,Request} requests sent by server, pass along to client(s)
These are like window/logMessage but actually display an alert or prompt to the user.

The exact logic of which clients need to see these is not super important to get right now. For these, it's less important than for window/logMessage that clients receive all previous requests when they join, but in theory that will happen for window/showMessage requests.