gitserver: Use a single temporary directory in ReposDir (#12288)
We would pollute the top-level ReposDir with temporary directories since we
want the directories to be on the same filesystem as the rest of the
repositories. If anything went wrong, the directories would not be
removed. However, we could just update TMP_DIR to be a directory inside of
ReposDir. We implement this behaviour for `ReposDir/.tmp` and additionally get
these benefits:

1. Clearing out `.tmp` on start-up
2. Set `$TMP_DIR` so any other temp files created (say by git) will also use
   the directory.
3. Easy to skip the directory when listing repositories.