graphql: Disallow git commands on disabled repository resolver (#12290)
Our backend happily returns disabled repositories, since we want to be able to
inspect them. But if you do a graphql query against a disabled repo asking for
its commit, we do not check if that is allowed.

Our gitserver API is not rich enough to prevent this, and it still contains code
to automatically start cloning (soon to be removed). In the meantime we can
prevent these mistakes at the graphql layer.

The main side-effect is the browser extension does a resolve-rev via graphql,
even if the repository is disabled. This ends up triggering a clone in
gitserver.