gitserver: Re-use cloneRepo in cleanup (#12321)
* gitserver: Factor clone goroutine into function

We want to make it possible to run cloneRepo in a blocking fashion. This will
allow us to share code with other parts more effectively.

* gitserver: Re-use repoClone in cleanup

* gitserver: Use correct repo name in cleanup

Updated the variable names to make this clearer. But previously we used the name
including the GIT_DIR. So we just needed to do filepath.Dir on the path.

* gitserver: Clean paths passed to NormalizeRepo

The most useful transformation this does is ensure we don't have a trailing `/`.
For example previously `github.com/foo/bar/.git` would be normalized to
`github.com/foo/bar/`.