web: use cxp-js for CXP support, pass along CXP state (Environment type)
The CXP controller is the interface between the React application and CXP extensions (servers or, soon, Web Workers). Its state is supplied by CONTROLLER.setEnvironment calls, so it is not a stateful service/component (except for internal state such as socket connection state, although it does not expose that to the outside world).