web,cxp: support extensions run via JS bundles, add sample bundled extension
This adds a new way of running extensions, in a Web Worker on the client. To do this, the extension specifies a URL where a JS bundle resides. This URL is imported in a Web Worker with `importScripts`. The bundle usually bundles in the `cxp` npm package and speaks full CXP over Web Worker `postMessage`.