Try to diagnose possible rate-limiting from github (#12328)
* Try to diagnose possible rate-limiting from github

It looks like the mysterious failures are *probably* us hitting
the "abuse rate limiting" feature on github, and that in turn
might be because repo-updater is hitting things more often with
the new scheduler, but it might also be for other reasons.

But the first thing would be to confirm/deny that this is what's
happening. To that end, two changes:

* In repo-updater's github code, check for the specific format
  documented for 403 errors.
* In github-proxy, dump contents of 4xx errors.

This is not a really good solution, but it might be suitable
enough for running on dogfood for a while to see whether we
can cause the failures to happen again and confirm that this
is the problem. A better fix requires deciding what to do about
getting such errors, but almost certainly one of the priorities
should be making sure we don't keep spamming things when getting
them, so probably caching them for a while?