rate-limiting in github-proxy, also fix rate-limiting (#12338)
* Move pkg/ratelimit out of repo-updater so it can be shared

Specifically, github-proxy has been gathering rate limits and
dumping them to Prometheus, but not actually *checking* them, and
I think it needs to check them to give us a failsafe to reduce
spamminess; github proxy should absolutely not be making
queries when it could easily know it's rate-limited.

* Check whether we're *after* a reset.

`t.Before(u)` => "t is before u". Thus, `time.Now().Before(c.reset)`
means "if the current time is not yet as late as the time at which
we expected a reset', which means that, if our most recent rate query
is recent enough that we should be trusting the data, we're
unconditionally resetting the rate limit right away. That's almost
certainly not what we want.

The test computations were subtly wrong, because they were testing the
values generated by the incorrect algorithm.

* Honor rate limits

Store rate limits for each resource type, and honor them in
github-proxy, not just elsewhere in the system. If we actually
run out of rate limit, this means no queries until it resets,
which is good, because *they wouldn't work*.

* document the github-proxy rate limiting fix