Pass on caching headers (#12363)
repo-updater already has an httpcache on its connections to github-proxy,
and httpcache is smart enough to handle things like If-Unmodified-Since,
Etag, and other headers that control caching behavior. Which means that
if we pass those on, we should be able to get github-proxy passing
them to github, and possibly getting back cache-friendly results
saying things like "yes, that's still good", which cause httpcache
to yield the same results as last time without us having to implement
anything else for it.