gitserver: Run reclone GC for every instance (#12361)
Previously we would only run reclone GC on sourcegraph.com. This was an
oversight, which meant busy repos would slow down over time. We have seen this
happen on dogfood a few times with the kubernetes repository.

Additionally included in this commit is some other refactorings and fixes in
code related to the cleanup task.