Fix search input not updating (#12371)
Fixes #12364, which was caused by a commit that added the `GlobalNavbar` component on the homepage.

Only the `SearchNavbarItem` component, which isn't rendered on the search homepage,  listened to route changes and skipped the first emitted observable, which caused the issue. So, this makes it so that it's parent, the `GlobalNavbar` component, is listening to route changes and is responsible for updating the query appropriately.