Refactoring of createEnableUpdateRepos behavior (#12369)
* Slow down repo-updater schedule

The fastest-time of 15 seconds was fast enough that active repos
were being updated about every 15-20 seconds, which was fairly spammy,
and probably unnecessary. Reduce frequency by adjusting the timing
for repos with updates in last two days a bit, to scale from 45
seconds to about 5 minutes, instead of 15 seconds to about 5 minutes.

(Note, this will still be more often than the old algorithm for
actively-updated repos.)

* Refactor createEnableUpdateRepos (part 1)

No caller of createEnableUpdateRepos ever used the slice argument.
Also, there's no reason every caller should be duplicating the
"make a channel" code. This is a precursor to improving the
functionality in more complex ways; specifically, we want to
duplicate the behavior we have for the configured-repository
list, where we detect items that cease showing up in the list.
That will require adding a way to distinguish which list is being
processed, but first I want to simplify the interface and calling
conventions a bit.

* Partially rework configuration/repolist thing

Refactoring the configUpdate code to be shareable with other code;
next step would be making createEnableUpdateRepos use the same
logic.

* Further refactoring of createEnableUpdateRepos

Add a "source" argument which can be used by the various code
host services to distinguish their repository lists. Switch
everything to using the improved/simpler logic, if we have
the new scheduler up.

* Update repo URLs when processing lists

It's possible for repo URLs to change, since URLs can contain
authentication tokens, which can change. Copy new URLs into the
repo list during updates of the list.