repo-updater: Do not purge new clones (#12411)
We want to prevent removing repositories which have just been cloned. This is to
allow users to manually place repositories into gitserver.

We add more information to gitservers repo info endpoint. We do not need to
worry about the order this is deployed since the information is just ignored if
we speak to an older gitserver.