improve confmerge functionality (#12206)
confmerge relies on the ability to represent the config values from
the config.json and config_local.json files in an internal structure
obtained from schema/schema.go. If you add an experimental flag,
and don't rebuild confmerge, it will *remove* that flag even if it's
present in either or both config. confmerge needs to be built from
the current code to work correctly. So build it unconditionally.

Also, have confmerge detect and warn about possible config file
problems, such as unsupported keys that will be dropped, so it's
easier to spot this.

Also fix silly script typo.