feat: remove experimental settings actions contributions
These are not needed now that we have commands, window/showMessageRequest, and configuration/update.

BREAKING CHANGE: Contributions may no longer specify settings actions (which were experimental). If an extension relies on these to update settings in response to the user clicking on a contribution's action, it must be updated to use some combination of workspace/executeCommand, window/showMessageRequest, and configuration/update.