reduce change-watching spam (#12440)
* improve handle-change diagnostics

* ensure failure from go-install.sh gets reported more visibly

* Batch compilation to reduce spamminess

If you make a change in pkg/*, we rebuild everything. If we rebuild
them all with separate commands, that means a lot of spam in the
logs. Recombine the go build commands, reducing it to one for
race-detector things and one for non-race-detector things.

This does result in a chance that changes which come in simultaneously
for two packages, one of which doesn't build, cause the other to
not get restarted. This is more tolerable than the spam, and
improved logging means you'll at least know the build didn't
happen.