Drop rate-limiting from github-proxy (#12504)
The rate-limiting implementation in GHP was racey, but also
it was wrong; it was assuming that rate limits were the same
for each "resource" globally, rather than per-token, so it was
wrong in cases with multiple tokens. But also, it probably shouldn't
have been doing rate-limiting at all. Or if it should, that needs
to be more carefully considered and planned out, and should probably
use a fundamentally different rate-limit algorithm.

The way this was returning rate limits (or rather, failing out
because of them and *not* returning them) broke computations of
rate limiting elsewhere, so let's just drop it and go back to
a state that we know worked except for the constant 403 errors,
which were addressed much more successfully by passing on cache-friendly
headers.

Partial revert of fb3c88d8c29f72d23c24e43fe84f80db3a82f88e.