Don't use ReposCreateIfNotExists without an actual reason  (#12436)
* Don't use ReposCreateIfNotExists without an actual reason

When I initially wrote this, I didn't quite understand the
relationship between ReposCreateIfNotExists and the rest of the
periodic update logic. Thus, every time through the update
loop, I was calling it but not really using the response.

The correct behavior is to call it when seeing a possibly-new
repo for the first time (as in, when one of the periodic config
checks finds new repos), and then to use the returned response
as guidance for what's *actually* enabled, rather than assuming
the default state. createEnableUpdateRepos was doing this
correctly, but the config watch case wasn't. But also, drop
the call that doesn't actually get used, that was just a
"let's copy the logic from the other updates and assume it's
reasonable" thing, it doesn't need to be there.

* Set "debouncing" interval to something small when processing manual updates

Now that the manual flag should be used only when external services
have explicitly requested an update, we should ensure that such explicit
requests get a shorter "debounce" time.

The interval sent to RequestRepoUpdate determines how long ago the
last fetch has to be before gitserver will actually do a fetch. Most
repos will have an interval of a minute or more, some will have an
interval of hours. We don't want that to override user-initiated things
like visits to a repo in the UI.