repo-updater: Simplify ping implementation (#12523)
The scheduler sleeps until nextDue before doing more work. The ping channel is
used to wake up the scheduler when there is new work to do which should be
done before nextDue. The channel was implemented with a ready variable to
avoid the blocking send which lead to deadlock. However, we can achieve the
same result by using a non-blocking send and a channel with buffer size 1.