docker management: use a proper --rm implementation (#12534)
docker management: use a proper --rm implementation

In my previous attempt to implement `--rm` (see 89e790a27589f48cc9f0dc24d604a7b5f0ba36ab)
I incorrectly thought that removing the container after we stopped it would be effectively
the same. This logic was based on the idea that `docker run` on a stopped/exited container
would effectively start it. However, this is NOT true. `docker run` does not start a
stopped/exited/killed container and it instead must be removed in order to reallocate the
container name.

See the detailed inline comment for further explanation.

Despite being verbose (for clarity) this is a low-risk change, and is important to get
into the release. It primarily affected only us (Felix and Isaac) because we sometimes
use `DEBUG_MANAGED_DOCKER=t` which is much more likely to kill the frontend process
non-gracefully, hence leaving language server containers in such a state. However, it
can happen with customers, too, for example by running `docker kill typescript` or by
language servers that crash or exit unusually.

Fixes #12241