langservers: warn if 'docker stop <langserver>' takes longer than 10s
We have a `Kill` flag because some language servers (e.g. Java) today
do not respond to `docker stop` and it makes us wait an additional 10s
before killing the container.

This adds a warning log for such cases so that we can know for certain
whether or not this is the case. This was useful when investigating
https://github.com/sourcegraph/sourcegraph/issues/12454 as it told me
the issue was _not_ this (as it had been in the past), but rather
something else. This also helps with developing language servers.