fix(web): reduce jitter when navigating to a new file
Previously, while the new blob was loading, the repo header contributions (eg "toggle line wrap", "toggle discussions", "toggle history panel") would be hidden. This causes UI jitter because when switching to a new file, they would disappear for however long it took to load (usually a few hundred msec to a few seconds).

This change makes those contributions remain. There is now a slight amount of jitter when switching to a new Markdown (rendered) file, but it is only on the leftmost contribution, so it does not look bad.